Added Key and changed rotation code
added normalization of the direction vector as it simplifies trigonometric calculations required for rotation. It ensures that the camera remains at the same distance from the ball after rotation, preventing any distortion in the rotation.
also added a key 'v' or 'V' which sets the camera position to the initial position the camera was in so you could toggle between Top view and initial view.